adds workManagerController's tests to SyncPresenterImpl
